fixed clearHint thunk and score now reads from 0 to 9
